added set concentration factor function
updated set concentration factor function; added the set concentration value function and setting of the new composition grid csv.

updated the setting of the new file and the setting of the concentration value and factor according to the wanted checkup files and bugs.

debugged the set_concentrations_factor function and the creation of new_grid so it actually works.